+++ lib/karafka/processing/schedulers/default.rb
@@ -24,6 +24,7 @@ module Karafka
         alias on_schedule_revocation on_schedule_consumption
         alias on_schedule_shutdown on_schedule_consumption
         alias on_schedule_idle on_schedule_consumption
+        alias on_schedule_eofed on_schedule_consumption
 
         # This scheduler does not have anything to manage as it is a pass through and has no state
         def on_manage