+++ spec/integrations/pro/consumption/eofed/vps_with_eof_spec.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+# When using virtual partitions with eof, each VP should receive its own `#eofed` execution.
+
+setup_karafka do |config|
+  config.kafka[:'enable.partition.eof'] = true
+  config.concurrency = 10
+end
+
+class Consumer < Karafka::BaseConsumer
+  def consume; end
+
+  def eofed
+    DT[:eofed] << object_id
+  end
+end
+
+draw_routes do
+  topic DT.topic do
+    consumer Consumer
+    virtual_partitions(
+      partitioner: lambda(&:raw_payload)
+    )
+  end
+end
+
+start_karafka_and_wait_until do
+  produce_many(DT.topic, DT.uuids(100))
+  sleep(5)
+
+  DT[:eofed].uniq.size >= 8
+end